+++ src/bdd/wisp/steps.py
@@ -1,11 +1,9 @@
-import request_generator as requestgen
-import session as session
+import src.utility.wisp.request_generator as requestgen
+import src.utility.wisp.session as session
 from behave import *
 
-import src.utility.steps_utils as steputils
-from src.utility import constants
-from src.utility import routes as router
-from src.utility import utils
+import src.utility.wisp.steps_utils as steputils
+from src.utility.wisp import utils, routes as router, constants
 
 
 @given(
