+++ src/utility/wisp/request_generator.py
@@ -4,8 +4,7 @@ import uuid
 
 import session as session
 
-from src.utility import constants
-from src.utility import utils
+from src.utility.wisp import utils, constants
 
 
 # ==============================================
